+++ demo/App.vue
@@ -1,5 +1,5 @@
 <template>
-    <div id="app" class="px-4">
+    <div id="app" class="px-4 fill-height">
         <Demo />
     </div>
 </template>
@@ -8,7 +8,7 @@
 import Demo from './Demo'
 
 export default {
-    name: 'app',
+    name: 'App',
     components: {
         Demo
     }
@@ -16,14 +16,12 @@ export default {
 </script>
 
 <style lang="scss">
-#demo-app {
+#app {
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
     text-align: left;
     color: #2c3e50;
-    padding-top: 60px;
     width: 100%;
-    overflow: hidden;
     max-width: 100%;
 }