+++ src/utils/assign.js
@@ -1,5 +1,5 @@
 export function assign(source, target) {
-    for (let a in source) {
+    for (const a in source) {
         if (target.hasOwnProperty(a)) {
             target[a] = source[a]
         }