+++ tests/unit/sidebar.spec.js
@@ -1,6 +1,6 @@
 import CompositionApi from '@vue/composition-api'
 import { shallowMount, createLocalVue } from '@vue/test-utils'
-import Sidebar from '../../src/components/Sidebar.vue'
+import Sidebar from '@/components/Sidebar.vue'
 const localVue = createLocalVue()
 localVue.use(CompositionApi)